+++ src/card.js
@@ -16,32 +16,7 @@ const evaluateGuess = (guess, correctAnswer) => {
   }
 }
 
-const createDeck = (allCards) => {
-  let deck = {
-    cards: allCards
-  }
-  return deck
-}
-
-const countCards = (allCards) => {
-  let deckLength = allCards.cards.length
-  return deckLength
-}
-
-const createRound = (deck) => {
-  let round = {
-    deck: deck,
-    currentCard: deck.cards[0],
-    turns: 0,
-    incorrectGuesses: []
-  }
-  return round
-}
-
 module.exports = {
   createCard,
   evaluateGuess,
-  createDeck,
-  countCards,
-  createRound,
 }
