+++ src/deck.js
@@ -1,0 +1,16 @@
+const createDeck = (allCards) => {
+  let deck = {
+    cards: allCards
+  }
+  return deck
+}
+
+const countCards = (allCards) => {
+  let deckLength = allCards.cards.length
+  return deckLength
+}
+
+module.exports = {
+  createDeck,
+  countCards,
+}
